+++ python/GafferUI/NameSwitchUI.py
@@ -183,9 +183,9 @@ class _InPlugValueWidget( GafferUI.PlugValueWidget ) :
 		GafferUI.PlugValueWidget.setReadOnly( self, readOnly )
 		self.__plugLayout.setReadOnly( readOnly )
 
-	def childPlugValueWidget( self, childPlug, lazy=True ) :
+	def childPlugValueWidget( self, childPlug ) :
 
-		return self.__plugLayout.plugValueWidget( childPlug, lazy )
+		return self.__plugLayout.plugValueWidget( childPlug )
 
 	def _updateFromPlug( self ) :
 
@@ -384,7 +384,7 @@ class _RowPlugValueWidget( GafferUI.PlugValueWidget ) :
 
 		return True
 
-	def childPlugValueWidget( self, childPlug, lazy=True ) :
+	def childPlugValueWidget( self, childPlug ) :
 
 		for w in self.__plugValueWidgets :
 			if w.getPlug().isSame( childPlug ) :